+++ core/classes/class.homepage.php
@@ -2,9 +2,35 @@
 
 class Homepage
 {
+    const PAGE_INDEX = 'index';
+    const PAGE_PHPINFO = 'phpinfo';
+    
+    private $page;
+    private $pageList = array(
+        self::PAGE_INDEX,
+        self::PAGE_PHPINFO,
+    );
+    
     public function __construct()
     {
         Util::logInitClass($this);
+        
+        $page = Util::cleanGetVar('p');
+        $this->page = !empty($page) && in_array($page, $this->pageList) ? $page : self::PAGE_INDEX;
+    }
+    
+    public function getPage()
+    {
+        return $this->page;
+    }
+    
+    public function getPageUrl($page)
+    {
+        $request = '';
+        if (!empty($page) && in_array($page, $this->pageList) && $page != self::PAGE_INDEX) {
+            $request = '/?p=' . $page;
+        }
+        return $this->getUrl() . $request;
     }
     
     public function getPath()
@@ -51,4 +77,6 @@ class Homepage
         $result .= PHP_EOL . '</Directory>';
         return file_put_contents($this->getAliasFilePath(), $result) !== false;
     }
+    
+
 }