+++ core/resources/homepage/tpls/hp.index.apache.php
@@ -84,3 +84,22 @@
   } ?>
   </div>
 </div>
+<div class="row">
+  <div class="col-lg-12">
+    <h3><i class="fa fa-windows"></i> <?php echo $neardLang->getValue(Lang::WINDOWS_HOSTS); ?> <small></small></h3>
+  </div>
+</div>
+<div class="row">
+  <div class="col-lg-12">
+  <?php foreach(Util::getHosts() as $host) {
+    if ($host['ip'] == '127.0.0.1') {
+      if ($host['enabled']) {
+        ?><div style="float:left;padding:3px;"><a class="btn btn-default" target="_blank" href="<?php echo 'http://' . $host['domain']; ?>"><span class="fa fa-check-square-o"></span> <?php echo $host['domain']; ?></a></div><?php
+      } else {
+        ?><div style="float:left;padding:3px;"><a class="btn btn-default" target="_blank" href="<?php echo 'http://' . $host['domain']; ?>"><span class="fa fa-square-o"></span> <?php echo $host['domain']; ?></a></div><?php
+      }
+    }
+    
+  } ?>
+  </div>
+</div>